requirements.txt
Flask==3.0.3
flask-cors==4.0.1

python-dotenv==1.0.1

google-api-python-client==2.154.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

openai==1.51.2
